+++ CHANGELOG.md
@@ -5,6 +5,13 @@
 ## Added 
 - ability to add multiple rooms/files to a page
 
+##
+## Changed
+- hide sign out button on page docspace
+- hide request name for anonymous
+- structure of tables with files (Name,Size,Type)
+- base theme in admin panel for docspace
+
 ## 2.0.0
 ## Added
 - support for public rooms